Renamed the GSTiling class to make it clear what it actually does.
