+++ jplag/src/main/java/jplag/strategy/RevisionComparisonStrategy.java
@@ -4,14 +4,14 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Vector;
 import jplag.JPlagComparison;
-import jplag.GSTiling;
+import jplag.GreedyStringTiling;
 import jplag.JPlagOptions;
 import jplag.JPlagResult;
 import jplag.Submission;
 
 public class RevisionComparisonStrategy extends AbstractComparisonStrategy {
 
-    public RevisionComparisonStrategy(JPlagOptions options, GSTiling gSTiling) {
+    public RevisionComparisonStrategy(JPlagOptions options, GreedyStringTiling gSTiling) {
         super(options, gSTiling);
     }
